+++ library/src/geometry/alias.rs
@@ -1,5 +1,5 @@
 ﻿use cgmath::Point3;
 use cgmath::Vector3;
 
-pub type Point = Point3<f32>;
-pub type Vector = Vector3<f32>;
+pub type Point = Point3<f64>;
+pub type Vector = Vector3<f64>;
