+++ library/src/geometry/epsilon.rs
@@ -1,1 +1,1 @@
-﻿pub(crate) const DEFAULT_EPSILON: f32 = 1e-7;
+﻿pub(crate) const DEFAULT_EPSILON: f64 = 1e-7;
